pyproject.toml
[project]
name = "linebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "gunicorn>=23.0.0",
    "line-bot-sdk>=3.17.1",
    "psycopg2>=2.9.10",
    "pypika>=0.48.9",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
]
